Modify Bpref implementation to work on short result lists (Bpref@x)
trec_eval Bpref implementation (see
http://icb.med.cornell.edu/wiki/index.php/BPrefTrecEval2006) does not behave
well if the number of results (x) is lower than the number of documents known
to be relevant to the topic according to the ground truth (R).

For example, we are inspecting only the first 20 documents (Bpref@20) from the
search engine for a particular topic. If we know more than 20 documents to be
relevant for the topic in the ground truth the original trec_eval
implementation of Bpref will not result in score 1.0 even if all the top 20
results being inspected are relevant.


git-svn-id: https://mir.fi.muni.cz/svn/mir/software/MIREVal@2266 95de04d9-2432-4d89-94e9-5886ee2f69d8